splitting more unit tests up
